+++ pkg/store/additive_columns_internal_test.go
@@ -32,6 +32,7 @@ var additiveColumnSources = map[int][]map[string]string{
 	23: {secretsSharedCols, triggerClaimOwnerCols},
 	24: {triggerWebhookDeliveryCols},
 	25: {triggerWebhookReplayKeyCols},
+	28: {nodesOrderCols},
 }
 
 func columnSpecMaps() []map[string]string {